news_crawler/requirements.txt
requests
beautifulsoup4
lxml
google-auth
google-auth-oauthlib
google-api-python-client
